nearing a first plugin running
after some refactoring the server will boot, bind the correct port,
and register plugins. It responds to a heartbeat api endpoint, but
breaks when submitting to the checkClient endpoint. Upgraded to the
es6 promises polyfill, as recommended by our previous promises lib,
but need to account for the API differences.