Framework complete. Plugin on the way.
the framework is working, and the proof-of-concept plugin for search
engine check is underway. there is a test folder with no tests, but
with a sample JSON blob of version 3 that can be fired at the server.

For testing I recommend getting your personal FHR packet, saving it
alongside the sample blob, and submitting it with:

    curl -X POST /
      -d @tests/payloads/my-data.json /
      --header "Content-Type:application/json" /
      0.0.0.0:8080/checkClient

Replacing the final line with your development address and port. If
you have trouble, enable the debug plugin in settings to verify that
the message is being received.